feat: Extract exercise logic to Calendar component
Refactors the calendar component to accept children,
moving exercise management logic from Calendar to Home
page.  This improves component separation and
maintainability.  No breaking changes.
